Make page components work again after themifying. Invalid values for .Params.component throw an error, and a missing value falls back to the content/block component.
Background: The process for validating if a component exists was based on a file path lookup. The Hugo file lookup function requires a file path relative to the root project, so looking for the layout in a theme requires explicitly looking in the theme's directory. There is no nice way to do this because the theme name cannot be retreived programmatically and their is no guarantee that the theme is in a directory with any one name.